Extend motif score calculation to any sequence lengths
Before, the motif score calculation hard codes the sequence length as 2*motif_len-1, which is the case for SNP.

Now, the motif score calculation takes the sequence length as an input argument, so that it supports any sequence length as long as it is longer than motif_len. This supports Indel cases.